requirements.txt
opencv-contrib-python
numpy
tensorflow
keras
python-socketio
Flask
eventlet
